fix removeMeta

+++ app/src/main/java/com/main/hooker/hooker/utils/State.java
@@ -70,7 +70,7 @@ public class State {
         try {
             IDAO<MetaItem> dao = getMetaDao();
             List<MetaItem> list = dao.findByCondition("key = '" + key + "'");
-            if (list == null || list.size() > 0) {
+            if (list != null && list.size() > 0) {
                 MetaItem item = list.get(0);
                 dao.delete(item.getId());
             }
